+++ distribution-entities/src/main/java/ru/wkn/streams/Stream.java
@@ -4,5 +4,5 @@ import ru.wkn.distributions.Distribution;
 
 public abstract class Stream {
 
-    public abstract double[] initTimeIntervals(Distribution distribution, int timeCoefficient);
+    public abstract double[] initTimeIntervals(Distribution distribution, double timeCoefficient);
 }
